Very basic rendering on map load

+++ Sledge.Rendering/Scenes/Renderables/Vertex.cs
@@ -5,10 +5,10 @@ namespace Sledge.Rendering.Scenes.Renderables
     public class Vertex
     {
         public Vector3 Position { get; set; }
-        public decimal TextureU { get; set; }
-        public decimal TextureV { get; set; }
+        public float TextureU { get; set; }
+        public float TextureV { get; set; }
 
-        public Vertex(Vector3 position, decimal textureU, decimal textureV)
+        public Vertex(Vector3 position, float textureU, float textureV)
         {
             Position = position;
             TextureU = textureU;
